query, data get test

+++ src/main/java/com/example/mapstudy/vo/IntersectionVO.java
@@ -7,6 +7,6 @@ public class IntersectionVO {
     String areaName;
     double areaLat;
     double areaLon;
-    long traffic;
+    int traffic;
     double delay;
 }
